move specific strategy inside netlist too. Runs end-to-end
